backend/requirements.txt
# Milestone 1 - read-only viewer + auth + signed assets
fastapi>=0.109.0,<0.115
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
argon2-cffi>=23.1.0
python-multipart>=0.0.6
httpx>=0.26.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
Pillow>=10.0.0
# AWS (only required when STORAGE_BACKEND=s3 or AWS_SECRETS_ARN is set)
boto3>=1.34.0
# Production ASGI server (optional for local dev)
gunicorn>=22.0.0
# Metrics (Prometheus format)
prometheus-client>=0.19.0